+++ opentelemetry-sdk/src/opentelemetry/sdk/_metrics/_view_instrument_match.py
@@ -15,10 +15,12 @@
 
 from logging import getLogger
 from threading import Lock
-from typing import TYPE_CHECKING, Iterable
+from typing import TYPE_CHECKING, Dict, Iterable
 
 from opentelemetry.sdk._metrics.aggregation import (
+    _Aggregation,
     _convert_aggregation_temporality,
+    _PointVarT,
 )
 from opentelemetry.sdk._metrics.measurement import Measurement
 from opentelemetry.sdk._metrics.point import AggregationTemporality, Metric
@@ -41,8 +43,8 @@ class _ViewInstrumentMatch:
         self._view = view
         self._instrument = instrument
         self._sdk_config = sdk_config
-        self._attributes_aggregation = {}
-        self._attributes_previous_point = {}
+        self._attributes_aggregation: Dict[frozenset, _Aggregation] = {}
+        self._attributes_previous_point: Dict[frozenset, _PointVarT] = {}
         self._lock = Lock()
 
     # pylint: disable=protected-access
@@ -65,15 +67,11 @@ class _ViewInstrumentMatch:
         if attributes not in self._attributes_aggregation:
             with self._lock:
                 if attributes not in self._attributes_aggregation:
-                    if self._view._aggregation:
-                        aggregation = (
-                            self._view._aggregation._create_aggregation(
-                                self._instrument
-                            )
-                        )
-                    else:
-                        aggregation = self._instrument._default_aggregation
-                    self._attributes_aggregation[attributes] = aggregation
+                    self._attributes_aggregation[
+                        attributes
+                    ] = self._view._aggregation._create_aggregation(
+                        self._instrument
+                    )
 
         self._attributes_aggregation[attributes].aggregate(measurement)