+++ opentelemetry-sdk/src/opentelemetry/sdk/_metrics/aggregation.py
@@ -12,6 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""
+.. Explicitly document private _AggregationFactory
+.. autoclass:: _AggregationFactory
+"""
+
 from abc import ABC, abstractmethod
 from bisect import bisect_left
 from dataclasses import replace
@@ -22,18 +27,20 @@ from typing import Generic, List, Optional, Sequence, TypeVar
 
 from opentelemetry._metrics.instrument import (
     Asynchronous,
+    Counter,
+    Histogram,
     Instrument,
+    ObservableCounter,
+    ObservableGauge,
+    ObservableUpDownCounter,
     Synchronous,
+    UpDownCounter,
     _Monotonic,
 )
 from opentelemetry.sdk._metrics.measurement import Measurement
-from opentelemetry.sdk._metrics.point import (
-    AggregationTemporality,
-    Gauge,
-    Histogram,
-    PointT,
-    Sum,
-)
+from opentelemetry.sdk._metrics.point import AggregationTemporality, Gauge
+from opentelemetry.sdk._metrics.point import Histogram as HistogramPoint
+from opentelemetry.sdk._metrics.point import PointT, Sum
 from opentelemetry.util._time import _time_ns
 
 _PointVarT = TypeVar("_PointVarT", bound=PointT)
@@ -52,6 +59,74 @@ class _Aggregation(ABC, Generic[_PointVarT]):
     @abstractmethod
     def collect(self) -> Optional[_PointVarT]:
         pass
+
+
+class _DropAggregation(_Aggregation):
+    def aggregate(self, measurement: Measurement) -> None:
+        pass
+
+    def collect(self) -> Optional[_PointVarT]:
+        pass
+
+
+class _AggregationFactory(ABC):
+    @abstractmethod
+    def _create_aggregation(self, instrument: Instrument) -> _Aggregation:
+        """Creates an aggregation"""
+
+
+class DefaultAggregation(_AggregationFactory):
+    """
+    The default aggregation to be used in a `View`.
+
+    This aggregation will create an actual aggregation depending on the
+    instrument type, as specified next:
+
+    ========================= ====================================
+    Instrument                Aggregation
+    ========================= ====================================
+    `Counter`                 `SumAggregation`
+    `UpDownCounter`           `SumAggregation`
+    `ObservableCounter`       `SumAggregation`
+    `ObservableUpDownCounter` `SumAggregation`
+    `Histogram`               `ExplicitBucketHistogramAggregation`
+    `ObservableGauge`         `LastValueAggregation`
+    ========================= ====================================
+    """
+
+    def _create_aggregation(self, instrument: Instrument) -> _Aggregation:
+
+        # pylint: disable=too-many-return-statements
+        if isinstance(instrument, Counter):
+            return _SumAggregation(
+                instrument_is_monotonic=True,
+                instrument_temporality=AggregationTemporality.DELTA,
+            )
+        if isinstance(instrument, UpDownCounter):
+            return _SumAggregation(
+                instrument_is_monotonic=False,
+                instrument_temporality=AggregationTemporality.DELTA,
+            )
+
+        if isinstance(instrument, ObservableCounter):
+            return _SumAggregation(
+                instrument_is_monotonic=True,
+                instrument_temporality=AggregationTemporality.CUMULATIVE,
+            )
+
+        if isinstance(instrument, ObservableUpDownCounter):
+            return _SumAggregation(
+                instrument_is_monotonic=False,
+                instrument_temporality=AggregationTemporality.CUMULATIVE,
+            )
+
+        if isinstance(instrument, Histogram):
+            return _ExplicitBucketHistogramAggregation()
+
+        if isinstance(instrument, ObservableGauge):
+            return _LastValueAggregation()
+
+        raise Exception(f"Invalid instrument type {type(instrument)} found")
 
 
 class _SumAggregation(_Aggregation[Sum]):
@@ -141,7 +216,7 @@ class _LastValueAggregation(_Aggregation[Gauge]):
         )
 
 
-class _ExplicitBucketHistogramAggregation(_Aggregation[Histogram]):
+class _ExplicitBucketHistogramAggregation(_Aggregation[HistogramPoint]):
     def __init__(
         self,
         boundaries: Sequence[float] = (
@@ -182,7 +257,7 @@ class _ExplicitBucketHistogramAggregation(_Aggregation[Histogram]):
 
         self._bucket_counts[bisect_left(self._boundaries, value)] += 1
 
-    def collect(self) -> Histogram:
+    def collect(self) -> HistogramPoint:
         """
         Atomically return a point for the current value of the metric.
         """
@@ -191,17 +266,19 @@ class _ExplicitBucketHistogramAggregation(_Aggregation[Histogram]):
         with self._lock:
             value = self._bucket_counts
             start_time_unix_nano = self._start_time_unix_nano
+            histogram_sum = self._sum
 
             self._bucket_counts = self._get_empty_bucket_counts()
             self._start_time_unix_nano = now + 1
+            self._sum = 0
 
-        return Histogram(
+        return HistogramPoint(
             start_time_unix_nano=start_time_unix_nano,
             time_unix_nano=now,
             bucket_counts=tuple(value),
             explicit_bounds=self._boundaries,
             aggregation_temporality=AggregationTemporality.DELTA,
-            sum=self._sum,
+            sum=histogram_sum,
         )
 
 
@@ -226,8 +303,10 @@ def _convert_aggregation_temporality(
     if current_point_type is Gauge:
         return current_point
 
-    if previous_point is not None and type(previous_point) is not type(
-        current_point
+    if (
+        previous_point is not None
+        and current_point is not None
+        and type(previous_point) is not type(current_point)
     ):
         _logger.warning(
             "convert_aggregation_temporality called with mismatched "
@@ -280,7 +359,7 @@ def _convert_aggregation_temporality(
             is_monotonic=is_monotonic,
         )
 
-    if current_point_type is Histogram:
+    if current_point_type is HistogramPoint:
         if previous_point is None:
             return replace(
                 current_point, aggregation_temporality=aggregation_temporality
@@ -314,7 +393,7 @@ def _convert_aggregation_temporality(
                 )
             ]
 
-        return Histogram(
+        return HistogramPoint(
             start_time_unix_nano=start_time_unix_nano,
             time_unix_nano=current_point.time_unix_nano,
             bucket_counts=bucket_counts,
@@ -323,12 +402,6 @@ def _convert_aggregation_temporality(
             aggregation_temporality=aggregation_temporality,
         )
     return None
-
-
-class _AggregationFactory(ABC):
-    @abstractmethod
-    def _create_aggregation(self, instrument: Instrument) -> _Aggregation:
-        """Creates an aggregation"""
 
 
 class ExplicitBucketHistogramAggregation(_AggregationFactory):
@@ -376,3 +449,10 @@ class SumAggregation(_AggregationFactory):
 class LastValueAggregation(_AggregationFactory):
     def _create_aggregation(self, instrument: Instrument) -> _Aggregation:
         return _LastValueAggregation()
+
+
+class DropAggregation(_AggregationFactory):
+    """Using this aggregation will make all measurements be ignored."""
+
+    def _create_aggregation(self, instrument: Instrument) -> _Aggregation:
+        return _DropAggregation()