+++ opentelemetry-sdk/tests/metrics/test_aggregation.py
@@ -23,6 +23,7 @@ from unittest.mock import Mock
 from opentelemetry.sdk._metrics import instrument
 from opentelemetry.sdk._metrics.aggregation import (
     AggregationTemporality,
+    DefaultAggregation,
     ExplicitBucketHistogramAggregation,
     LastValueAggregation,
     SumAggregation,
@@ -31,8 +32,18 @@ from opentelemetry.sdk._metrics.aggregation import (
     _LastValueAggregation,
     _SumAggregation,
 )
+from opentelemetry.sdk._metrics.instrument import (
+    Counter,
+    Histogram,
+    ObservableCounter,
+    ObservableGauge,
+    ObservableUpDownCounter,
+    UpDownCounter,
+)
 from opentelemetry.sdk._metrics.measurement import Measurement
-from opentelemetry.sdk._metrics.point import Gauge, Histogram, Sum
+from opentelemetry.sdk._metrics.point import Gauge
+from opentelemetry.sdk._metrics.point import Histogram as HistogramPoint
+from opentelemetry.sdk._metrics.point import Sum
 from opentelemetry.util.types import Attributes
 
 
@@ -286,6 +297,7 @@ class TestExplicitBucketHistogramAggregation(TestCase):
         first_histogram = explicit_bucket_histogram_aggregation.collect()
 
         self.assertEqual(first_histogram.bucket_counts, (0, 1, 0, 0))
+        self.assertEqual(first_histogram.sum, 1)
 
         # CI fails the last assertion without this
         sleep(0.1)
@@ -294,6 +306,7 @@ class TestExplicitBucketHistogramAggregation(TestCase):
         second_histogram = explicit_bucket_histogram_aggregation.collect()
 
         self.assertEqual(second_histogram.bucket_counts, (0, 1, 0, 0))
+        self.assertEqual(second_histogram.sum, 1)
 
         self.assertGreater(
             second_histogram.time_unix_nano, first_histogram.time_unix_nano
@@ -346,6 +359,17 @@ class TestConvertAggregationTemporality(TestCase):
                 ),
                 current_point,
             )
+
+        with self.assertRaises(AssertionError):
+            with self.assertLogs(level=WARNING):
+                self.assertIs(
+                    _convert_aggregation_temporality(
+                        Gauge(time_unix_nano=0, value=0),
+                        None,
+                        AggregationTemporality.DELTA,
+                    ),
+                    current_point,
+                )
 
     def test_current_point_sum_previous_point_none(self):
 
@@ -599,7 +623,7 @@ class TestConvertAggregationTemporality(TestCase):
 class TestHistogramConvertAggregationTemporality(TestCase):
     def test_previous_point_none(self):
 
-        current_point = Histogram(
+        current_point = HistogramPoint(
             start_time_unix_nano=0,
             time_unix_nano=1,
             bucket_counts=[0, 2, 1, 2, 0],
@@ -623,7 +647,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
         with self.assertRaises(Exception):
 
             _convert_aggregation_temporality(
-                Histogram(
+                HistogramPoint(
                     start_time_unix_nano=0,
                     time_unix_nano=1,
                     bucket_counts=[0, 2, 1, 2, 0],
@@ -631,7 +655,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
                     sum=70,
                     aggregation_temporality=AggregationTemporality.DELTA,
                 ),
-                Histogram(
+                HistogramPoint(
                     start_time_unix_nano=1,
                     time_unix_nano=2,
                     bucket_counts=[0, 1, 3, 0, 0],
@@ -643,7 +667,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
             ),
 
     def test_same_aggregation_temporality_cumulative(self):
-        current_point = Histogram(
+        current_point = HistogramPoint(
             start_time_unix_nano=0,
             time_unix_nano=2,
             bucket_counts=[0, 3, 4, 2, 0],
@@ -653,7 +677,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
         )
         self.assertEqual(
             _convert_aggregation_temporality(
-                Histogram(
+                HistogramPoint(
                     start_time_unix_nano=0,
                     time_unix_nano=1,
                     bucket_counts=[0, 2, 1, 2, 0],
@@ -668,7 +692,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
         )
 
     def test_same_aggregation_temporality_delta(self):
-        current_point = Histogram(
+        current_point = HistogramPoint(
             start_time_unix_nano=1,
             time_unix_nano=2,
             bucket_counts=[0, 1, 3, 0, 0],
@@ -679,7 +703,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
 
         self.assertEqual(
             _convert_aggregation_temporality(
-                Histogram(
+                HistogramPoint(
                     start_time_unix_nano=0,
                     time_unix_nano=2,
                     bucket_counts=[0, 3, 4, 2, 0],
@@ -694,7 +718,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
         )
 
     def test_aggregation_temporality_to_cumulative(self):
-        current_point = Histogram(
+        current_point = HistogramPoint(
             start_time_unix_nano=1,
             time_unix_nano=2,
             bucket_counts=[0, 1, 3, 0, 0],
@@ -705,7 +729,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
 
         self.assertEqual(
             _convert_aggregation_temporality(
-                Histogram(
+                HistogramPoint(
                     start_time_unix_nano=0,
                     time_unix_nano=1,
                     bucket_counts=[0, 2, 1, 2, 0],
@@ -716,7 +740,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
                 current_point,
                 AggregationTemporality.CUMULATIVE,
             ),
-            Histogram(
+            HistogramPoint(
                 start_time_unix_nano=0,
                 time_unix_nano=2,
                 bucket_counts=[0, 3, 4, 2, 0],
@@ -727,7 +751,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
         )
 
     def test_aggregation_temporality_to_delta(self):
-        current_point = Histogram(
+        current_point = HistogramPoint(
             start_time_unix_nano=0,
             time_unix_nano=2,
             bucket_counts=[0, 3, 4, 2, 0],
@@ -738,7 +762,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
 
         self.assertEqual(
             _convert_aggregation_temporality(
-                Histogram(
+                HistogramPoint(
                     start_time_unix_nano=0,
                     time_unix_nano=1,
                     bucket_counts=[0, 2, 1, 2, 0],
@@ -749,7 +773,7 @@ class TestHistogramConvertAggregationTemporality(TestCase):
                 current_point,
                 AggregationTemporality.DELTA,
             ),
-            Histogram(
+            HistogramPoint(
                 start_time_unix_nano=1,
                 time_unix_nano=2,
                 bucket_counts=[0, 1, 3, 0, 0],
@@ -815,3 +839,80 @@ class TestAggregationFactory(TestCase):
         self.assertIsInstance(aggregation, _LastValueAggregation)
         aggregation2 = factory._create_aggregation(counter)
         self.assertNotEqual(aggregation, aggregation2)
+
+
+class TestDefaultAggregation(TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.default_aggregation = DefaultAggregation()
+
+    def test_counter(self):
+
+        aggregation = self.default_aggregation._create_aggregation(
+            Counter(Mock(), Mock(), Mock())
+        )
+        self.assertIsInstance(aggregation, _SumAggregation)
+        self.assertTrue(aggregation._instrument_is_monotonic)
+        self.assertEqual(
+            aggregation._instrument_temporality, AggregationTemporality.DELTA
+        )
+
+    def test_up_down_counter(self):
+
+        aggregation = self.default_aggregation._create_aggregation(
+            UpDownCounter(Mock(), Mock(), Mock())
+        )
+        self.assertIsInstance(aggregation, _SumAggregation)
+        self.assertFalse(aggregation._instrument_is_monotonic)
+        self.assertEqual(
+            aggregation._instrument_temporality, AggregationTemporality.DELTA
+        )
+
+    def test_observable_counter(self):
+
+        aggregation = self.default_aggregation._create_aggregation(
+            ObservableCounter(Mock(), Mock(), Mock(), Mock())
+        )
+        self.assertIsInstance(aggregation, _SumAggregation)
+        self.assertTrue(aggregation._instrument_is_monotonic)
+        self.assertEqual(
+            aggregation._instrument_temporality,
+            AggregationTemporality.CUMULATIVE,
+        )
+
+    def test_observable_up_down_counter(self):
+
+        aggregation = self.default_aggregation._create_aggregation(
+            ObservableUpDownCounter(Mock(), Mock(), Mock(), Mock())
+        )
+        self.assertIsInstance(aggregation, _SumAggregation)
+        self.assertFalse(aggregation._instrument_is_monotonic)
+        self.assertEqual(
+            aggregation._instrument_temporality,
+            AggregationTemporality.CUMULATIVE,
+        )
+
+    def test_histogram(self):
+
+        aggregation = self.default_aggregation._create_aggregation(
+            Histogram(
+                Mock(),
+                Mock(),
+                Mock(),
+                Mock(),
+                Mock(),
+            )
+        )
+        self.assertIsInstance(aggregation, _ExplicitBucketHistogramAggregation)
+
+    def test_observable_gauge(self):
+
+        aggregation = self.default_aggregation._create_aggregation(
+            ObservableGauge(
+                Mock(),
+                Mock(),
+                Mock(),
+                Mock(),
+            )
+        )
+        self.assertIsInstance(aggregation, _LastValueAggregation)