+++ opentelemetry-sdk/tests/metrics/test_metric_reader_storage.py
@@ -14,12 +14,15 @@
 
 from unittest.mock import Mock, patch
 
+from opentelemetry.sdk._metrics.aggregation import DropAggregation
+from opentelemetry.sdk._metrics.instrument import Counter
 from opentelemetry.sdk._metrics.measurement import Measurement
 from opentelemetry.sdk._metrics.metric_reader_storage import (
     MetricReaderStorage,
 )
 from opentelemetry.sdk._metrics.point import AggregationTemporality
 from opentelemetry.sdk._metrics.sdk_configuration import SdkConfiguration
+from opentelemetry.sdk._metrics.view import View
 from opentelemetry.test.concurrency_test import ConcurrencyTestBase, MockFunc
 
 
@@ -53,7 +56,6 @@ class TestMetricReaderStorage(ConcurrencyTestBase):
                 resource=Mock(),
                 metric_readers=(),
                 views=(view1, view2),
-                enable_default_view=True,
             )
         )
 
@@ -98,7 +100,6 @@ class TestMetricReaderStorage(ConcurrencyTestBase):
                 resource=Mock(),
                 metric_readers=(),
                 views=(view1, view2),
-                enable_default_view=True,
             )
         )
 
@@ -146,7 +147,6 @@ class TestMetricReaderStorage(ConcurrencyTestBase):
                 resource=Mock(),
                 metric_readers=(),
                 views=(view1,),
-                enable_default_view=True,
             )
         )
 
@@ -172,7 +172,6 @@ class TestMetricReaderStorage(ConcurrencyTestBase):
                 resource=Mock(),
                 metric_readers=(),
                 views=(),
-                enable_default_view=True,
             )
         )
 
@@ -189,31 +188,22 @@ class TestMetricReaderStorage(ConcurrencyTestBase):
         storage.consume_measurement(Measurement(1, instrument2))
         self.assertEqual(len(MockViewInstrumentMatch.call_args_list), 1)
 
-    @patch(
-        "opentelemetry.sdk._metrics.metric_reader_storage._ViewInstrumentMatch"
-    )
-    def test_default_view_disabled(self, MockViewInstrumentMatch: Mock):
-        """Instruments should not be matched with default views when disabled"""
-        instrument1 = Mock(name="instrument1")
-        instrument2 = Mock(name="instrument2")
-        storage = MetricReaderStorage(
+    def test_drop_aggregation(self):
+
+        counter = Counter("name", Mock(), Mock())
+        metric_reader_storage = MetricReaderStorage(
             SdkConfiguration(
                 resource=Mock(),
                 metric_readers=(),
-                views=(),
-                enable_default_view=False,
+                views=(
+                    View(
+                        instrument_name="name", aggregation=DropAggregation()
+                    ),
+                ),
             )
         )
+        metric_reader_storage.consume_measurement(Measurement(1, counter))
 
-        storage.consume_measurement(Measurement(1, instrument1))
         self.assertEqual(
-            len(MockViewInstrumentMatch.call_args_list),
-            0,
-            MockViewInstrumentMatch.mock_calls,
+            [], metric_reader_storage.collect(AggregationTemporality.DELTA)
         )
-        storage.consume_measurement(Measurement(1, instrument1))
-        self.assertEqual(len(MockViewInstrumentMatch.call_args_list), 0)
-
-        MockViewInstrumentMatch.call_args_list.clear()
-        storage.consume_measurement(Measurement(1, instrument2))
-        self.assertEqual(len(MockViewInstrumentMatch.call_args_list), 0)