+++ CHANGELOG.md
@@ -7,11 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v1.10.0-0.29b0...HEAD)
 
+- Add default aggregation
+  ([#2543](https://github.com/open-telemetry/opentelemetry-python/pull/2543))
 - Fix incorrect installation of some exporter “convenience” packages into
   “site-packages/src”
   ([#2525](https://github.com/open-telemetry/opentelemetry-python/pull/2525))
+- Capture exception information as part of log attributes
+  ([#2531](https://github.com/open-telemetry/opentelemetry-python/pull/2531))
 - Change OTLPHandler to LoggingHandler
   ([#2528](https://github.com/open-telemetry/opentelemetry-python/pull/2528))
+- Fix delta histogram sum not being reset on collection
+  ([#2533](https://github.com/open-telemetry/opentelemetry-python/pull/2533))
+- Add InMemoryMetricReader to metrics SDK
+  ([#2540](https://github.com/open-telemetry/opentelemetry-python/pull/2540))
+- Drop the usage of name field from log model in OTLP
+  ([#2565](https://github.com/open-telemetry/opentelemetry-python/pull/2565))
+- Update opentelemetry-proto to v0.15.0
+  ([#2566](https://github.com/open-telemetry/opentelemetry-python/pull/2566))
+- Remove `enable_default_view` option from sdk MeterProvider
+  ([#2547](https://github.com/open-telemetry/opentelemetry-python/pull/2547))
 
 ## [1.10.0-0.29b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.10.0-0.29b0) - 2022-03-10